+++ packages/web3-providers/src/entry-types.ts
@@ -22,6 +22,7 @@ export * from './types/Price.js'
 export * from './types/RSS3.js'
 export * from './types/ScamWarning.js'
 export * from './types/Security.js'
+export * from './types/SimpleHash.js'
 export * from './types/TokenIcon.js'
 export * from './types/TokenList.js'
 export * from './types/Trending.js'
